+++ example/lib/main.dart
@@ -102,7 +102,7 @@ class PopupDialog extends StatelessWidget {
             "COOL",
             style: TextStyle(color: Colors.white, fontSize: 20),
           ),
-          onPressed: () => Navigator.pop(context),
+          onPressed: () => _onCustomAnimationAlertPressed(context),
           width: 120,
         )
       ],
@@ -144,16 +144,25 @@ class PopupDialog extends StatelessWidget {
   _onAlertWithStylePressed(context) {
     // Reusable alert style
     var alertStyle = AlertStyle(
+      animationType: AnimationType.fromTop,
+      isCloseButton: false,
+      isOverlayTapDismiss: false,
+      descStyle: TextStyle(fontWeight: FontWeight.bold),
+      animationDuration: Duration(milliseconds: 400),
+      alertBorder: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(0.0),
+        side: BorderSide(
+          color: Colors.grey,
+        ),
+      ),
+      titleStyle: TextStyle(
+        color: Colors.red,
+      ),
+      constraints: BoxConstraints.expand(width: 300),
       //First to chars "55" represents transparency of color
-      overlayColor: Color(0x55000000)
+      overlayColor: Color(0x55000000),
+      alertElevation: 0,
     );
-
-    Alert(
-      style: alertStyle,
-      context: context,
-      title: "RFLUTTER ALERT",
-      desc: "Flutter is more awesome with RFlutter Alert.",
-    ).show();
 
     // Alert dialog using custom alert style
     Alert(